+++ lib/truffle/socket/truffle.rb
@@ -148,19 +148,6 @@ module Truffle
       0
     end
 
-    def self.family_for_sockaddr_in(sockaddr)
-      case sockaddr.bytesize
-      when Foreign::SockaddrIn6.size
-        ::Socket::AF_INET6
-      when Foreign::SockaddrIn.size
-        ::Socket::AF_INET
-      # UNIX socket addresses can have a variable size as sometimes any trailing
-      # null bytes are stripped (e.g. when calling UNIXServer#getsockname).
-      else
-        ::Socket::AF_UNIX
-      end
-    end
-
     def self.constant_pairs
       # Truffle: no need to filter here since only defined constants are in the config
       vals = {}